requirements.txt
Flask==2.0.1
gunicorn==20.1.0
jyserver==0.0.5
keras==2.6.0
matplotlib==3.4.3
mediapipe==0.9.0.1
numpy==1.19.5
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
Pillow==8.2.0
Pillow==9.5.0
tensorflow==2.6.0
